Moved command processing to its own class (CommandProcessor.java)
